Move `arrays` utilities into `utilities/common`
