+++ src/errors/ApolloError.ts
@@ -1,6 +1,6 @@
 import { GraphQLError } from 'graphql';
 
-import { isNonEmptyArray } from '../util/arrays';
+import { isNonEmptyArray } from '../utilities/common/arrays';
 
 export function isApolloError(err: Error): err is ApolloError {
   return err.hasOwnProperty('graphQLErrors');
